initial idea for shared tests
